requirements.txt
anyio==3.6.1
asgiref==3.5.0
certifi==2022.9.24
charset-normalizer==2.1.1
dadata==21.10.1
Django==4.0.2
djangorestframework==3.14.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
Pillow==9.0.1
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pytz==2022.4
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.12
